requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-nvidia-ai-endpoints==0.0.11
langchain-community==0.0.24
python-dotenv==1.0.0
chromadb==0.4.22
pydantic==2.6.1

# Docling and its dependencies
docling>=1.0.0
pillow>=10.0.0


streamlit
python-dotenv
langchain-nvidia-ai-endpoints
langchain-text-splitters
langchain-community
chromadb
pypdf
